tvOS target is buildable
