+++ MC Stats.xcodeproj/xcuserdata/topscrech.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -12,27 +12,27 @@
 		<key>MCStatusIntentsFramework.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>5</integer>
+			<integer>6</integer>
 		</dict>
 		<key>Widgets watchOS.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>4</integer>
+			<integer>5</integer>
 		</dict>
 		<key>Widgets.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>3</integer>
+			<integer>4</integer>
 		</dict>
 		<key>tvOS.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>7</integer>
+			<integer>2</integer>
 		</dict>
 		<key>visionOS.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
-			<integer>2</integer>
+			<integer>3</integer>
 		</dict>
 		<key>watchOS.xcscheme_^#shared#^_</key>
 		<dict>
@@ -43,6 +43,11 @@
 	<key>SuppressBuildableAutocreation</key>
 	<dict>
 		<key>F35E1A7C2D6501A500718CB7</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>F3C879BC2D65720500840745</key>
 		<dict>
 			<key>primary</key>
 			<true/>
