+++ MCStatus/Localizable.xcstrings
@@ -114,6 +114,9 @@
     "Got it!" : {
 
     },
+    "Hello, world!" : {
+
+    },
     "Help support the development of free, adless, open source apps" : {
 
     },
