+++ MCStatus/Views/FAQ/FAQ.swift
@@ -1,0 +1,12 @@
+import Foundation
+
+struct FAQ: Identifiable {
+    let id = UUID()
+    let question: String
+    let answer: String
+    
+    init(_ question: String, answer: String) {
+        self.question = question
+        self.answer = answer
+    }
+}
