+++ MCStatus/Views/Settings/ShortcutsGuideView.swift
@@ -20,9 +20,10 @@ struct ShortcutsGuideView: View {
         Form {
             // Overview Section
             Section {
+#if !os(tvOS)
                 ShortcutsLink()
                     .shortcutsLinkStyle(.automaticOutline)
-                
+#endif
                 Text("With Shortcuts, you can quickly check the status of your Minecraft servers without even opening the app. Use the built in shortcuts to show the current status and player count, or create custom shortcuts using the server's online status, current player count, player sample, and more")
                     .padding(.vertical, 5)
             } header: {
@@ -48,6 +49,8 @@ struct ShortcutsGuideView: View {
             }
         }
         .navigationTitle("Shortcuts")
+#if !os(tvOS)
         .background(Color(.systemGroupedBackground))
+#endif
     }
 }
